pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icsforge"
dynamic = ["version"]
description = "OT/ICS Cybersecurity Coverage Validation Framework — MITRE ATT&CK for ICS aligned"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    {name = "ICSForge Contributors"},
]
keywords = ["ics", "ot", "scada", "cybersecurity", "mitre-attack", "pcap", "modbus", "dnp3", "s7comm", "iec104", "opcua", "enip", "profinet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "flask>=3.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.2",
    "ruff>=0.4",
]
replay = [
    "scapy>=2.5",
]
kafka = [
    "kafka-python>=2.0",
]

[project.urls]
Homepage = "https://www.icsforge.nl"
Repository = "https://github.com/ICSforge/ICSforge"
Issues = "https://github.com/ICSforge/ICSforge/issues"
Documentation = "https://www.icsforge.nl"

[project.scripts]
icsforge = "icsforge.cli:main"
icsforge-web = "icsforge.web.app:main"
icsforge-receiver = "icsforge.receiver.receiver:main"

[tool.setuptools.dynamic]
version = {attr = "icsforge.__version__"}

[tool.setuptools.packages.find]
include = ["icsforge*"]

[tool.setuptools.package-data]
icsforge = [
    "scenarios/*.yml",
    "data/*.json",
    "profiles/*.yml",
    "campaigns/*.yml",
    "receiver/*.yml",
    "web/templates/*.html",
    "web/static/**/*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --timeout=30 -p no:cov"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "network: marks tests requiring network access",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter
